requirements.txt
django==1.11.4
gunicorn==19.7.1
scaleway-sdk==1.5.0